+++ rest/src/main/kotlin/com/gitlab/kordlib/rest/route/Route.kt
@@ -7,7 +7,6 @@ import kotlinx.serialization.*
 import kotlinx.serialization.builtins.ListSerializer
 import kotlinx.serialization.builtins.nullable
 import kotlinx.serialization.builtins.serializer
-import kotlinx.serialization.internal.*
 import com.gitlab.kordlib.common.entity.DiscordEmoji as EmojiEntity
 
 sealed class Route<T>(
@@ -232,7 +231,7 @@ sealed class Route<T>(
         : Route<List<InviteResponse>>(HttpMethod.Get, "/guilds/$GuildId/invites", ListSerializer(InviteResponse.serializer()))
 
     object GuildIntegrationGet
-        : Route<List<DiscordGuildIntegrations>>(HttpMethod.Get, "/guilds/$GuildId/integrations", ListSerializer(DiscordGuildIntegrations.serializer()))
+        : Route<List<IntegrationResponse>>(HttpMethod.Get, "/guilds/$GuildId/integrations", ListSerializer(IntegrationResponse.serializer()))
 
     object GuildIntegrationPost
         : Route<Unit>(HttpMethod.Post, "/guilds/$GuildId/integrations", NoStrategy)